+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Added
+- Constant Wait strategy
+
 ## [1.0.0] - 2016-08-29
 ### Added
 - Initial implementation